Implementação desfazer deleção via snackbar - PR19

+++ tafeito-ai/src/components/Task/index.tsx
@@ -9,12 +9,13 @@ import ListItemText from "@mui/material/ListItemText";
 import Checkbox from "@mui/material/Checkbox";
 
 import TaskActions from "../TaskActions";
+import { closeSnackbar } from "notistack";
 import { useGlobalContext } from '../../utils/global';
 
 import { TaskProps } from "./Task";
 import DeleteTaskDialog from "../DeleteTaskDialog";
 import { url_update_task, url_finish_task, url_reopen_task, url_estimate_task } from "../../utils/api";
-import { useSnackbar } from "notistack";
+import { useSnackbar, SnackbarKey } from "notistack";
 import TaskTags from "../TaskTags";
 
 const Task = (props:TaskProps) => {
@@ -23,7 +24,16 @@ const Task = (props:TaskProps) => {
         onTaskChange
     } = props;
 
-    const { setIsEditingTask, setRefetchTaskStatus, refetchTaskStatus, isLoading, setIsLoading } = useGlobalContext();
+    const { 
+      setIsEditingTask, 
+      setRefetchTaskStatus, 
+      refetchTaskStatus, 
+      setIsLoading, 
+      softDeletedTasks, 
+      setSoftDeletedTasks, 
+      softDeletedTasksRef 
+    } = useGlobalContext();
+
     const [ _error, setError ] = useState<null | string>(null);
 
     const [ openedDialog, setOpenedDialog ] = useState(false);
@@ -80,14 +90,46 @@ const Task = (props:TaskProps) => {
         setChecked(newChecked);
     };
 
+    const action = (snackbarId: SnackbarKey) => (
+      <>
+        <button
+          onClick={() => {
+            const filteredValues = softDeletedTasks.filter(x => x !== task?.id);
+            setSoftDeletedTasks(filteredValues);
+            closeSnackbar(snackbarId);
+          }}
+        >Undo</button>
+      </>
+    );
+
     const deleteTask = async () => {
       setIsLoading(true);
       const taskId = task?.id ?? -1;
       const custom_task_url = url_update_task.replace(":id", taskId.toString());
       try {
-        await api.delete(custom_task_url);
+        //await api.delete(custom_task_url);
+        const newSoftDeletedTasks = softDeletedTasks.concat(task?.id);
+        setSoftDeletedTasks(newSoftDeletedTasks);
+
         setError(null);
-        enqueueSnackbar("Tarefa deletada!", { variant: "success" });
+        //enqueueSnackbar("Tarefa deletada!", { variant: "success" });
+
+        enqueueSnackbar("Tarefa deletada!", {
+          variant: "success",
+          action: action,
+          onExited: async () => {
+  
+            if(softDeletedTasksRef?.current?.includes(task?.id)) {
+              await api.delete(custom_task_url);
+            }
+  
+            //o id da task está na lista de tarefas deletadas?
+            //se sim, deletar task
+            //se não, não fazer nada
+            setRefetchTaskStatus(refetchTaskStatus + 1);
+          },
+        });
+  
         setRefetchTaskStatus(refetchTaskStatus + 1);
         setIsLoading(false);
       } catch (err) {
@@ -126,9 +168,7 @@ const Task = (props:TaskProps) => {
         secondaryAction={
         <TaskActions 
           deleteTask={() => {
-            if (openedDialog === false) {
-              setOpenedDialog(true);
-            }
+            deleteTask();
           }}
           editTask={() => {
             onTaskChange(task.id);
